+++ src/main/java/qastudio/backend/domain/auth/service/EmailCommandService.java
@@ -8,8 +8,5 @@ import qastudio.backend.domain.auth.dto.response.EmailResponse;
 import java.io.UnsupportedEncodingException;
 
 public interface EmailCommandService {
-    void createCode();
-    MimeMessage createEmailForm(String email) throws MessagingException, UnsupportedEncodingException;
-    EmailResponse sendEmail(EmailRequest emailRequest);
-    String setContext(String code);
+
 }
