+++ src/main/java/qastudio/backend/domain/auth/service/EmailCommandServiceImpl.java
@@ -23,71 +23,4 @@ import java.util.Random;
 @RequiredArgsConstructor
 public class EmailCommandServiceImpl implements EmailCommandService {
 
-    private final JavaMailSender emailSender;
-    private final AccountTableRepository accountTableRepository;
-    private final EmailQueryServiceImpl emailQueryService;
-    private final SpringTemplateEngine templateEngine;
-    private String randomCode;
-
-    // 랜덤 코드 8자 생성
-    public void createCode() {
-        Random random = new Random();
-        StringBuffer key = new StringBuffer();
-
-        for(int i = 0; i < 8; i++) { // 8자리
-            int index = random.nextInt(3);
-
-            switch (index) {
-                case 0 : // 소문자
-                    key.append((char) ((int)random.nextInt(26) + 97));
-                    break;
-                case 1: // 대문자
-                    key.append((char) ((int)random.nextInt(26) + 65));
-                    break;
-                case 2: // 0-9 숫자
-                    key.append(random.nextInt(9));
-                    break;
-            }
-        }
-        randomCode = key.toString();
-    }
-
-    // 이메일 양식 작성
-    public MimeMessage createEmailForm(String email) throws MessagingException, UnsupportedEncodingException {
-        createCode();
-        String setFrom = "qastudio7@gmail.com"; // email-config에 설정한 자신의 이메일 주소
-        String toEmail = email; // 받는 사람
-        String title = "QASTUDIO 회원가입 인증 번호입니다."; // 제목
-
-        MimeMessage message = emailSender.createMimeMessage();
-        message.addRecipients(MimeMessage.RecipientType.TO, toEmail); // 보낼 이메일 설정
-        message.setSubject(title);
-        message.setFrom(setFrom);
-        message.setText(setContext(randomCode), "utf-8", "html");
-
-        return message;
-    }
-
-    // 이메일 전송
-    public EmailResponse sendEmail(EmailRequest emailRequest) throws BadRequestException {
-        // 이메일 중복 검사
-        emailQueryService.checkEmailDuplication(emailRequest);
-
-        try {
-            MimeMessage emailForm = createEmailForm(emailRequest.getEmail());
-            emailSender.send(emailForm);
-            return EmailResponse.builder()
-                    .authCode(randomCode)
-                    .build();
-        } catch (UnsupportedEncodingException | MessagingException e){
-            throw new BadRequestException(ErrorStatus.EMAIL_VERIFICATION_SEND_FAILED);
-        }
-    }
-
-    // context 설정
-    public String setContext(String code) {
-        Context context = new Context();
-        context.setVariable("code", code);
-        return templateEngine.process("mail", context);
-    }
 }
